wip. Re-factoring card edit logic
