+++ src/assets/css/SingleCard.css
@@ -1,0 +1,47 @@
+.cards__card {
+  /* position: relative; */
+  display: flex;
+  margin: 5px 0 0 0;
+  border-bottom: 1px solid #b4b9c5;
+  border-radius: 4px;
+  background-color: #ffffff;
+  cursor: pointer;
+  overflow: hidden;
+}
+
+.cards__name {
+  border: none;
+  padding: 10px;
+  width: 100%;
+  word-wrap: break-word;
+}
+
+.cards__card:hover {
+  background-color: #f3f4f6;
+}
+
+.cards__action-icons {
+  display: flex;
+  font-size: 0.7rem;
+}
+
+.cards__action-icon {
+  color: #77787b;
+  display: inline;
+  cursor:crosshair;
+  padding: 10px;
+  border-radius: 2px;
+  /* visibility: hidden; */
+}
+
+.cards__action-icon:hover {
+  /* visibility: visible; */
+  color: #4f5053;
+  background-color: #e3e3e7;
+}
+
+.cards__delete-btn:hover,
+.cards__edit-btn:hover {
+  cursor: pointer;
+  border-radius: 4px;
+}
